refactor: encapsulate addition logic
